Group grid opts into hook
